# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: AssetFlow Enterprise Asset & Resource Management System API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /users:
    get:
      operationId: listUsers
      tags: [users]
      summary: List all users
      parameters:
        - name: department
          in: query
          schema:
            type: integer
        - name: role
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/User"
  /users/{id}/role:
    patch:
      operationId: updateUserRole
      tags: [users]
      summary: Update user role (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserRoleInput"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /users/{id}/status:
    patch:
      operationId: updateUserStatus
      tags: [users]
      summary: Update user status (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserStatusInput"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
...
  /transfer-requests:
    get:
      operationId: listTransferRequests
      tags: [transfers]
      summary: List transfer requests
      parameters:
        - name: status
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of transfer requests
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TransferRequest"
    post:
      operationId: createTransferRequest
      tags: [transfers]
      summary: Request asset transfer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TransferRequestInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TransferRequest"
  /transfer-requests/{id}/approve:
    patch:
      operationId: approveTransferRequest
      tags: [transfers]
      summary: Approve a transfer request
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Approved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TransferRequest"
  /transfer-requests/{id}/reject:
    patch:
      operationId: rejectTransferRequest
      tags: [transfers]
      summary: Reject a transfer request
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Rejected
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TransferRequest"
...
components:
  schemas:
...
    User:
      type: object
      required: [id, name, email, role, status, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        role:
          type: string
        departmentId:
          type: ["integer", "null"]
        departmentName:
          type: ["string", "null"]
        status:
          type: string
        createdAt:
          type: string
        updatedAt:
          type: string
    UserRoleInput:
      type: object
      required: [role]
      properties:
        role:
          type: string
    UserStatusInput:
      type: object
      required: [status]
      properties:
        status:
          type: string
...
    TransferRequest:
      type: object
      required: [id, assetId, toEmployeeId, status, requestedAt]
      properties:
        id:
          type: integer
        assetId:
          type: integer
        assetName:
          type: ["string", "null"]
        assetTag:
          type: ["string", "null"]
        fromEmployeeId:
          type: ["integer", "null"]
        fromEmployeeName:
          type: ["string", "null"]
        toEmployeeId:
          type: integer
        toEmployeeName:
          type: ["string", "null"]
        reason:
          type: ["string", "null"]
        status:
          type: string
        requestedBy:
          type: ["integer", "null"]
        approvedBy:
          type: ["integer", "null"]
        requestedAt:
          type: string
        resolvedAt:
          type: ["string", "null"]
    TransferRequestInput:
      type: object
      required: [assetId, toEmployeeId]
      properties:
        assetId:
          type: integer
        toEmployeeId:
          type: integer
        reason:
          type: ["string", "null"]
...